requirements.txt
absl-py==2.2.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
audioread==3.0.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.2.2
fastapi==0.115.12
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
idna==3.10
imbalanced-learn==0.12.4
joblib==1.4.2
keras==2.15.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
ml-dtypes==0.2.0
msgpack==1.1.0
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
onnx==1.17.0
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.15.2
seaborn==0.12.2
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.46.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.15.0
termcolor==2.5.0
tf_keras==2.15.1
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wrapt==1.14.1
xgboost==2.0.3
